+++ helper/terraformtype/helper/customdiff/funcs.go
@@ -1,0 +1,13 @@
+package customdiff
+
+const (
+	FuncNameAll            = `All`
+	FuncNameComputedIf     = `ComputedIf`
+	FuncNameForceNewIf     = `ForceNewIf`
+	FuncNameIf             = `If`
+	FuncNameIfValue        = `IfValue`
+	FuncNameIfValueChange  = `IfValueChange`
+	FuncNameSequence       = `Sequence`
+	FuncNameValidateChange = `ValidateChange`
+	FuncNameValidateValue  = `ValidateValue`
+)